# src/main/resources/static/swagger-apis/openapi.yaml
...
openapi: 3.0.3
swagger: 2.0
info:
  title: Evently API
  description: This API manages events data for its React client Evently App.
  termsOfService: urn:tos
  contact: {}
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: base url
"host": http://localhost:8080
"basePath": /api/v1
...
paths:
...
  "/events/{id}/directions":
    get:
      tags:
      - Direction
      summary: Get direction by event id
      operationId: getByEventIdUsingGET
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        description: id
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Direction"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
      - Direction
      summary: Create event direction
      operationId: createUsingPOST
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        description: id
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/DirectionReq"
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Message"
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
      - Direction
      summary: Delete event direction
      operationId: deleteUsingDELETE
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        description: id
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Message"
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden        
...
  "/events/{id}/participations":
    get:
      tags:
      - Participation
      summary: Get participations by event id
      operationId: getByEventIdUsingGET_1
      parameters:
      - name: id
        in: path
        description: id
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/ParticipationRes"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
      - Participation
      summary: Create a participation (join)
      operationId: createUsingPOST_2
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        description: id
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Message"
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
  "/events/{id}/requirements":
    get:
      tags:
      - Requirement
      summary: Get requirements by event id
      operationId: getByEventIdUsingGET_2
      parameters:
      - name: id
        in: path
        description: id
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Requirement"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
      - Requirement
      summary: Create event requirement
      operationId: createUsingPOST_3
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        description: id
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RequirementReq"
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Message"
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
      - Requirement
      summary: Delete event requirement
      operationId: deleteUsingDELETE_4
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        description: id
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RequirementReq"
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Message"
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  "/events/{id}/webUrls":
    get:
      tags:
      - Web url
      summary: Get by event id
      operationId: getByEventIdUsingGET_3
      parameters:
      - name: id
        in: path
        description: id
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/WebUrl"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
      - Web url
      summary: Create event url
      operationId: createUsingPOST_4
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        description: id
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/WebUrlReq"
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Message"
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
      - Web url
      summary: Delete event url
      operationId: deleteUsingDELETE_6
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        description: id
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Message"
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  "/webUrls":
    get:
      tags:
      - Web url
      summary: Get all
      operationId: getAllUsingGET_6
      security:
      - bearer: []
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/WebUrl"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/webUrls/{id}":
    get:
      tags:
      - Web url
      summary: Get a single url by its id
      operationId: getByIdUsingGET_4
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        description: id
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/WebUrl"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
components:
  securitySchemes:
    bearer:
      type: apiKey
      name: Authorization
      in: header
      description: Token de autenticación
  schemas:
    Direction:
      title: Direction
      type: object
      properties:
        building:
          type: string
        city:
          type: string
        country:
          type: string
        door:
          type: string
        province:
          type: string
        street:
          type: string
    DirectionReq:
      title: DirectionReq
      required:
      - building
      - city
      - country
      - province
      - street
      type: object
      properties:
        building:
          maxLength: 12
          minLength: 1
          type: string
        city:
          maxLength: 300
          minLength: 3
          type: string
        country:
          maxLength: 300
          minLength: 3
          type: string
        door:
          type: string
        province:
          maxLength: 300
          minLength: 3
          type: string
        street:
          maxLength: 300
          minLength: 3
          type: string
...
    Message:
      title: Message
      type: object
      properties:
        message:
          type: string
    NestedEventRes:
      title: NestedEventRes
      type: object
      properties:
        date:
          type: string
        description:
          type: string
        hour:
          type: string
        id:
          type: integer
          format: int64
        images:
          type: array
          items:
            type: string
        location:
          type: string
        participantsCount:
          type: integer
          format: int32
        publisher:
          "$ref": "#/components/schemas/NestedUser"
        title:
          type: string
        type:
          type: string
    NestedUser:
      title: NestedUser
      type: object
      properties:
        avatar:
          type: string
        completeName:
          type: string
        username:
          type: string
...
    ParticipationRes:
      title: ParticipationRes
      type: object
      properties:
        event:
          "$ref": "#/components/schemas/NestedEventRes"
        id:
          type: integer
          format: int64
        participant:
          "$ref": "#/components/schemas/NestedUser"
...
    Requirement:
      title: Requirement
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    RequirementReq:
      title: RequirementReq
      required:
      - name
      type: object
      properties:
        name:
          maxLength: 50
          minLength: 2
          type: string
...
    WebUrl:
      title: WebUrl
      type: object
      properties:
        id:
          type: integer
          format: int64
        url:
          type: string
    WebUrlReq:
      title: WebUrlReq
      required:
      - url
      type: object
      properties:
        url:
          maxLength: 2147483647
          minLength: 7
          type: string